+++ src/test/java/lotto/domain/MatchResultTest.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.domain;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -12,10 +12,8 @@ public class MatchResultTest {
     @Test
     @DisplayName("수익률 계산 : 수익률 = {(당첨금액*개수)의 합 / 구입금액} * 100 ")
     void 수익률_계산() {
-        MatchResult matchResult = new MatchResult(Arrays.asList(1,0,0,0));
+        MatchResult matchResult = new MatchResult(Arrays.asList(0,0,0,1,0,0,0));
 
-        double returnRate = matchResult.calculateReturnRate(14000);
-
-        assertThat(returnRate).isEqualTo(0.35);
+        assertThat(matchResult.calculateReturnRate(14000)).isEqualTo(0.35);
     }
 }
