+++ src/test/java/lotto/util/LottoUtilTest.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.util;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
